+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bug: Fix variable declarations.
 - Bug: Add invalid service check.
 - Bug: Don't check group membership for * zone during apply of Masquerade and Redirect.
+- Bug : Improve expand_item.
 - Todo : Translate new features.
 - Todo : Fix backup.cgi restore upload.