+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Bug: Don't check group membership for * zone during apply of Masquerade and Redirect.
 - Bug : Improve expand_item.
 - Bug : Rework LoadServices.
+- Bug : Enable strict mode.
 - Todo : Translate new features.
 - Todo : Fix backup.cgi restore upload.