+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Bug : Enable strict mode.
 - Bug : Add invalid use of MAC address check.
 - Bug : Add invalid nDPI service check.
+- Services : Add Wireguard on port 51820 service.
 - Todo : Translate new features.
 - Todo : Fix backup.cgi restore upload.